move screenshot code to its own screenshot module
