+++ frontend/src/ts/test/test-ui.ts
@@ -1,13 +1,11 @@
 import * as Notifications from "../elements/notifications";
 import * as ThemeColors from "../elements/theme-colors";
 import Config, * as UpdateConfig from "../config";
-import * as DB from "../db";
 import * as TestWords from "./test-words";
 import * as TestInput from "./test-input";
 import * as CustomText from "./custom-text";
 import * as Caret from "./caret";
 import * as OutOfFocus from "./out-of-focus";
-import * as Replay from "./replay";
 import * as Misc from "../utils/misc";
 import * as Strings from "../utils/strings";
 import * as JSONData from "../utils/json-data";
@@ -17,28 +15,17 @@ import * as SlowTimer from "../states/slow-timer";
 import * as CompositionState from "../states/composition";
 import * as ConfigEvent from "../observables/config-event";
 import * as Hangul from "hangul-js";
-import { format } from "date-fns/format";
-import { isAuthenticated } from "../firebase";
 import { debounce } from "throttle-debounce";
 import * as ResultWordHighlight from "../elements/result-word-highlight";
 import * as ActivePage from "../states/active-page";
 import Format from "../utils/format";
-import * as Loader from "../elements/loader";
-import { getHtmlByUserFlags } from "../controllers/user-flag-controller";
 import {
   TimerColor,
   TimerOpacity,
 } from "@monkeytype/contracts/schemas/configs";
 import { convertRemToPixels } from "../utils/numbers";
-import {
-  findSingleActiveFunboxWithFunction,
-  getActiveFunboxesWithFunction,
-} from "./funbox/list";
+import { findSingleActiveFunboxWithFunction } from "./funbox/list";
 import * as TestState from "./test-state";
-
-async function gethtml2canvas(): Promise<typeof import("html2canvas").default> {
-  return (await import("html2canvas")).default;
-}
 
 function createHintsHtml(
   incorrectLtrIndices: number[][],
@@ -625,236 +612,6 @@ export function colorful(tc: boolean): void {
   } else {
     $("#words").removeClass("colorfulMode");
   }
-}
-
-let firefoxClipboardNotificatoinShown = false;
-
-/**
- * Prepares UI, generates screenshot canvas using html2canvas, and reverts UI changes.
- * Returns the generated canvas element or null on failure.
- * Handles its own loader and basic error notifications for canvas generation.
- */
-async function generateScreenshotCanvas(): Promise<HTMLCanvasElement | null> {
-  Loader.show();
-  let revealReplay = false;
-  let revertCookie = false;
-
-  if (!$("#resultReplay").hasClass("hidden")) {
-    revealReplay = true;
-    Replay.pauseReplay();
-  }
-  if (
-    Misc.isElementVisible("#cookiesModal") ||
-    document.contains(document.querySelector("#cookiesModal"))
-  ) {
-    revertCookie = true;
-  }
-
-  // Define revert function locally or ensure it's accessible if defined elsewhere
-  const revertScreenshot = (): void => {
-    Loader.hide();
-    $("#ad-result-wrapper").removeClass("hidden");
-    $("#ad-result-small-wrapper").removeClass("hidden");
-    $("#testConfig").removeClass("hidden");
-    $(".pageTest .screenshotSpacer").remove();
-    $("#notificationCenter").removeClass("hidden");
-    $("#commandLineMobileButton").removeClass("hidden");
-    $(".pageTest .ssWatermark").addClass("hidden");
-    $(".pageTest .ssWatermark").text("monkeytype.com"); // Reset watermark text
-    $(".pageTest .buttons").removeClass("hidden");
-    $("noscript").removeClass("hidden");
-    $("#nocss").removeClass("hidden");
-    $("header, footer").removeClass("invisible");
-    $("#result").removeClass("noBalloons");
-    $(".wordInputHighlight").removeClass("hidden");
-    $(".highlightContainer").removeClass("hidden");
-    if (revertCookie) $("#cookiesModal").removeClass("hidden");
-    if (revealReplay) $("#resultReplay").removeClass("hidden");
-    if (!isAuthenticated()) {
-      $(".pageTest .loginTip").removeClass("hidden");
-    }
-    (document.querySelector("html") as HTMLElement).style.scrollBehavior =
-      "smooth";
-    for (const fb of getActiveFunboxesWithFunction("applyGlobalCSS")) {
-      fb.functions.applyGlobalCSS();
-    }
-  };
-
-  // --- UI Preparation ---
-  const dateNow = new Date(Date.now());
-  $("#resultReplay").addClass("hidden");
-  $(".pageTest .ssWatermark").removeClass("hidden");
-
-  const snapshot = DB.getSnapshot();
-  const ssWatermark = [format(dateNow, "dd MMM yyyy HH:mm"), "monkeytype.com"];
-  if (snapshot?.name !== undefined) {
-    const userText = `${snapshot?.name}${getHtmlByUserFlags(snapshot, {
-      iconsOnly: true,
-    })}`;
-    ssWatermark.unshift(userText);
-  }
-  $(".pageTest .ssWatermark").html(
-    ssWatermark
-      .map((el) => `<span>${el}</span>`)
-      .join("<span class='pipe'>|</span>")
-  );
-  $(".pageTest .buttons").addClass("hidden");
-  $("#notificationCenter").addClass("hidden");
-  $("#commandLineMobileButton").addClass("hidden");
-  $(".pageTest .loginTip").addClass("hidden");
-  $("noscript").addClass("hidden");
-  $("#nocss").addClass("hidden");
-  $("#ad-result-wrapper").addClass("hidden");
-  $("#ad-result-small-wrapper").addClass("hidden");
-  $("#testConfig").addClass("hidden");
-  // Ensure spacer is removed before adding a new one if function is called rapidly
-  $(".pageTest .screenshotSpacer").remove();
-  $(".page.pageTest").prepend("<div class='screenshotSpacer'></div>");
-  $("header, footer").addClass("invisible");
-  $("#result").addClass("noBalloons");
-  $(".wordInputHighlight").addClass("hidden");
-  $(".highlightContainer").addClass("hidden");
-  if (revertCookie) $("#cookiesModal").addClass("hidden");
-
-  for (const fb of getActiveFunboxesWithFunction("clearGlobal")) {
-    fb.functions.clearGlobal();
-  }
-
-  (document.querySelector("html") as HTMLElement).style.scrollBehavior = "auto";
-  window.scrollTo({ top: 0, behavior: "instant" as ScrollBehavior }); // Use instant scroll
-
-  // --- Target Element Calculation ---
-  const src = $("#result .wrapper");
-  if (!src.length) {
-    console.error("Result wrapper not found for screenshot");
-    Notifications.add("Screenshot target element not found", -1);
-    revertScreenshot();
-    return null;
-  }
-  // Ensure offset calculations happen *after* potential layout shifts from UI prep
-  await new Promise((resolve) => setTimeout(resolve, 50)); // Small delay for render updates
-
-  const sourceX = src.offset()?.left ?? 0;
-  const sourceY = src.offset()?.top ?? 0;
-  const sourceWidth = src.outerWidth(true) as number;
-  const sourceHeight = src.outerHeight(true) as number;
-
-  // --- Canvas Generation ---
-  try {
-    const paddingX = convertRemToPixels(2);
-    const paddingY = convertRemToPixels(2);
-
-    const canvas = await (
-      await gethtml2canvas()
-    )(document.body, {
-      backgroundColor: await ThemeColors.get("bg"),
-      width: sourceWidth + paddingX * 2,
-      height: sourceHeight + paddingY * 2,
-      x: sourceX - paddingX,
-      y: sourceY - paddingY,
-      logging: false, // Suppress html2canvas logs in console
-      useCORS: true, // May be needed if user flags/icons are external
-    });
-
-    revertScreenshot(); // Revert UI *after* canvas is successfully generated
-    return canvas;
-  } catch (e) {
-    Notifications.add(
-      Misc.createErrorMessage(e, "Error creating screenshot canvas"),
-      -1
-    );
-    revertScreenshot(); // Ensure UI is reverted on error
-    return null;
-  }
-}
-
-/**
- * Generates screenshot and attempts to copy it to the clipboard.
- * Falls back to opening in a new tab if clipboard access fails.
- * Handles notifications related to the copy action.
- * (This function should be used by the 'copy' command or the original button)
- */
-export async function screenshot(): Promise<void> {
-  const canvas = await generateScreenshotCanvas();
-  if (!canvas) {
-    // Error notification handled by generateScreenshotCanvas
-    return;
-  }
-
-  canvas.toBlob(async (blob) => {
-    if (!blob) {
-      Notifications.add("Failed to generate image data (blob is null)", -1);
-      return;
-    }
-    try {
-      // Attempt to copy using ClipboardItem API
-      const clipItem = new ClipboardItem(
-        Object.defineProperty({}, blob.type, {
-          value: blob,
-          enumerable: true,
-        })
-      );
-      await navigator.clipboard.write([clipItem]);
-      Notifications.add("Copied screenshot to clipboard", 1, { duration: 2 });
-    } catch (e) {
-      // Handle clipboard write error
-      console.error("Error saving image to clipboard", e);
-
-      // Firefox specific message (only show once)
-      if (
-        navigator.userAgent.toLowerCase().includes("firefox") &&
-        !firefoxClipboardNotificatoinShown
-      ) {
-        firefoxClipboardNotificatoinShown = true;
-        Notifications.add(
-          "On Firefox you can enable the asyncClipboard.clipboardItem permission in about:config to enable copying straight to the clipboard",
-          0,
-          { duration: 10 }
-        );
-      }
-
-      // General fallback notification and action
-      Notifications.add(
-        "Could not copy screenshot to clipboard. Opening in new tab instead (make sure popups are allowed)",
-        0,
-        { duration: 5 }
-      );
-      try {
-        // Fallback: Open blob in a new tab
-        const blobUrl = URL.createObjectURL(blob);
-        window.open(blobUrl);
-        // No need to revoke URL immediately as the new tab needs it.
-        // Browser usually handles cleanup when tab is closed or navigated away.
-      } catch (openError) {
-        Notifications.add("Failed to open screenshot in new tab", -1);
-        console.error("Error opening blob URL:", openError);
-      }
-    }
-  });
-}
-
-/**
- * Generates screenshot canvas and returns the image data as a Blob.
- * Handles notifications for canvas/blob generation errors.
- * (This function is intended to be used by the 'download' command)
- */
-export async function getScreenshotBlob(): Promise<Blob | null> {
-  const canvas = await generateScreenshotCanvas();
-  if (!canvas) {
-    // Notification already handled by generateScreenshotCanvas
-    return null;
-  }
-
-  return new Promise((resolve) => {
-    canvas.toBlob((blob) => {
-      if (!blob) {
-        Notifications.add("Failed to convert canvas to Blob for download", -1);
-        resolve(null);
-      } else {
-        resolve(blob); // Return the generated blob
-      }
-    }, "image/png"); // Explicitly request PNG format
-  });
 }
 
 export async function updateActiveWordLetters(
@@ -1788,10 +1545,6 @@ function updateLiveStatsColor(value: TimerColor): void {
     $("#liveStatsMini").addClass("timerText");
   }
 }
-
-$(".pageTest").on("click", "#saveScreenshotButton", () => {
-  void screenshot();
-});
 
 $(".pageTest #copyWordsListButton").on("click", async () => {
   let words;
